PhoneLinux-Desktop/requirements.txt
click==7.1.2
grpcio==1.34.0
grpcio-tools==1.34.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
protobuf==3.14.0
psycopg2==2.8.6
PySide2==5.15.1
python-environ==0.4.54
shiboken2==5.15.1
six==1.15.0
Werkzeug==1.0.1
